pokemon.service.spec tests
